Improvements for DTOs and converters structure
