+++ src/main/java/org/ohdsi/webapi/cohortdefinition/converter/CohortMetadataDTOToCohortDefinitionConverter.java
@@ -1,0 +1,8 @@
+package org.ohdsi.webapi.cohortdefinition.converter;
+
+import org.ohdsi.webapi.cohortdefinition.dto.CohortMetadataDTO;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CohortMetadataDTOToCohortDefinitionConverter extends BaseCohortDTOToCohortDefinitionConverter<CohortMetadataDTO> {
+}
